+++ extensions/androidx-paging3/src/commonMain/kotlin/app/cash/sqldelight/paging3/OffsetQueryPagingSource.kt
@@ -15,14 +15,7 @@
  */
 package app.cash.sqldelight.paging3
 
-import app.cash.paging.PagingSourceLoadParams
-import app.cash.paging.PagingSourceLoadParamsAppend
-import app.cash.paging.PagingSourceLoadParamsPrepend
-import app.cash.paging.PagingSourceLoadParamsRefresh
-import app.cash.paging.PagingSourceLoadResult
-import app.cash.paging.PagingSourceLoadResultInvalid
-import app.cash.paging.PagingSourceLoadResultPage
-import app.cash.paging.PagingState
+import androidx.paging.PagingState
 import app.cash.sqldelight.Query
 import app.cash.sqldelight.SuspendingTransacter
 import app.cash.sqldelight.Transacter
@@ -42,26 +35,25 @@ internal class OffsetQueryPagingSource<RowType : Any>(
   override val jumpingSupported get() = true
 
   override suspend fun load(
-    params: PagingSourceLoadParams<Int>,
-  ): PagingSourceLoadResult<Int, RowType> = withContext(context) {
+    params: LoadParams<Int>,
+  ): LoadResult<Int, RowType> = withContext(context) {
     val key = params.key ?: initialOffset
     val limit = when (params) {
-      is PagingSourceLoadParamsPrepend<*> -> minOf(key, params.loadSize)
+      is LoadParams.Prepend -> minOf(key, params.loadSize)
       else -> params.loadSize
     }
-    val getPagingSourceLoadResult: TransactionCallbacks.() -> PagingSourceLoadResultPage<Int, RowType> = {
+    val getPagingSourceLoadResult: TransactionCallbacks.() -> LoadResult.Page<Int, RowType> = {
       val count = countQuery.executeAsOne()
       val offset = when (params) {
-        is PagingSourceLoadParamsPrepend<*> -> maxOf(0, key - params.loadSize)
-        is PagingSourceLoadParamsAppend<*> -> key
-        is PagingSourceLoadParamsRefresh<*> -> if (key >= count - params.loadSize) maxOf(0, count - params.loadSize) else key
-        else -> error("Unknown PagingSourceLoadParams ${params::class}")
+        is LoadParams.Prepend -> maxOf(0, key - params.loadSize)
+        is LoadParams.Append -> key
+        is LoadParams.Refresh -> if (key >= count - params.loadSize) maxOf(0, count - params.loadSize) else key
       }
       val data = queryProvider(limit, offset)
         .also { currentQuery = it }
         .executeAsList()
       val nextPosToLoad = offset + data.size
-      PagingSourceLoadResultPage(
+      LoadResult.Page(
         data = data,
         prevKey = offset.takeIf { it > 0 && data.isNotEmpty() },
         nextKey = nextPosToLoad.takeIf { data.isNotEmpty() && data.size >= limit && it < count },
@@ -73,7 +65,7 @@ internal class OffsetQueryPagingSource<RowType : Any>(
       is Transacter -> transacter.transactionWithResult(bodyWithReturn = getPagingSourceLoadResult)
       is SuspendingTransacter -> transacter.transactionWithResult(bodyWithReturn = getPagingSourceLoadResult)
     }
-    (if (invalid) PagingSourceLoadResultInvalid<Int, RowType>() else loadResult) as PagingSourceLoadResult<Int, RowType>
+    (if (invalid) LoadResult.Invalid() else loadResult)
   }
 
   override fun getRefreshKey(state: PagingState<Int, RowType>) = state.anchorPosition?.let { maxOf(0, it - (state.config.initialLoadSize / 2)) }