+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 - [Compiler] Allow for underscores in package names. Before underscores have been sanitized which resulted in unexpected behaviours (#6027 by @BierDav)
+- [Paging Extension] Switch to AndroidX Paging (#5910 by @jeffdgr8)
 
 ## [2.2.1] - 2025-11-13
 [2.2.1]: https://github.com/sqldelight/sqldelight/releases/tag/2.2.1